pyproject.toml
[project]
name = "nonebot-plugin-template"
version = "0.0.1"
description = "插件描述"
readme = "README.md"
authors = [{ name = "fllesser", email = "fllessive@gmail.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
  "httpx>=0.27.0,<1.0.0",
  "nonebot2>=2.4.0,<3.0.0",
  "nonebot-adapter-onebot>=2.4.6,<3.0.0",
  # "nonebot-plugin-uninfo>=0.6.10,<1.0.0",
  # "nonebot-plugin-alconna>=0.55.1,<1.0.0",
  # "nonebot-plugin-localstore>=0.7.3,<1.0.0",
  # "nonebot-plugin-apscheduler>=0.5.0,<1.0.0",
]

[dependency-groups]
dev = [
  "nonebot2[fastapi]>=2.4.0,<3.0.0",
  "nonebot-adapter-onebot>=2.4.6,<3.0.0",
]

lint = ["ruff"]

[tool.nonebot]
plugins = ["nonebot_plugin_aivoice"]
